introduce n5 interfaces that are responsible for version checks
