+++ src/main/java/org/janelia/saalfeldlab/n5/N5VersionedWriter.java
@@ -1,0 +1,11 @@
+package org.janelia.saalfeldlab.n5;
+
+import java.io.IOException;
+
+public interface N5VersionedWriter extends N5VersionedReader, N5Writer {
+
+	public default void setVersion() throws IOException {
+
+		setAttribute("/", VERSION_KEY, VERSION);
+	}
+}
